♻️ refactor(user): consolidate error mapping functions across routes
- Create shared error-mapper.ts with mapUserError function
- Create shared errors.ts with UserNotFoundError, EmailAlreadyTakenError, UsernameAlreadyTakenError
- Update all routes to use the shared error mapper
- Fix outer catch blocks in update-progress and get-profile routes
- Update test imports to use shared error classes
- Reduces code duplication and improves maintainability

AI Review: Consolidate error mapping functions
