+++ apps/api/src/features/user/update-progress/route.ts
@@ -8,48 +8,8 @@ import type { LoggerVariables } from '@api/middleware/logger';
 import type { Clock } from '@api/shared/clock';
 import { Hono } from 'hono';
 import type { IUserRepository } from '../domain/repositories/IUserRepository';
-import { UserNotFoundError, updateProgressHandler } from './handler';
-
-// Error mapper for update progress errors
-function mapUpdateProgressError(error: Error): { status: number; body: object } {
-  if (error.name === 'ValidationError') {
-    return {
-      status: 400,
-      body: {
-        success: false,
-        error: {
-          code: 'VALIDATION_ERROR',
-          message: error.message,
-        },
-      },
-    };
-  }
-
-  if (error instanceof UserNotFoundError) {
-    return {
-      status: 404,
-      body: {
-        success: false,
-        error: {
-          code: 'USER_NOT_FOUND',
-          message: error.message,
-        },
-      },
-    };
-  }
-
-  // Default error response
-  return {
-    status: 500,
-    body: {
-      success: false,
-      error: {
-        code: 'REPOSITORY_ERROR',
-        message: 'Internal server error',
-      },
-    },
-  };
-}
+import { mapUserError } from '../shared/error-mapper';
+import { updateProgressHandler } from './handler';
 
 // Define context variables for this route
 type UpdateProgressVariables = {
@@ -96,7 +56,7 @@ export const updateProgressRoute = new Hono<{
         errorMessage: error.message,
       });
 
-      const { status, body: errorBody } = mapUpdateProgressError(error);
+      const { status, body: errorBody } = mapUserError(error);
       return c.json(errorBody, status as SupportedStatusCode);
     }
 
@@ -118,15 +78,9 @@ export const updateProgressRoute = new Hono<{
       error: error instanceof Error ? error.message : String(error),
       stack: error instanceof Error ? error.stack : undefined,
     });
-    return c.json(
-      {
-        success: false,
-        error: {
-          code: 'REPOSITORY_ERROR',
-          message: 'Internal server error',
-        },
-      },
-      500
+    const { status, body: errorBody } = mapUserError(
+      error instanceof Error ? error : new Error('Unknown error')
     );
+    return c.json(errorBody, status as SupportedStatusCode);
   }
 });
